// src/types/MainListType.ts
export type MainListType = {
  addr1?: string;
  addr2?: string;
  areacode?: string;
  cat1?: string;
  cat2?: string;
  cat3?: string;
  contentid?: string;
  contenttypeid?: string;
  createdtime?: string;
  firstimage?: string;
  firstimage2?: string;
  mapx?: string;
  mapy?: string;
  mlevel?: string;
  modifiedtime?: string;
  readcount?: number;
  sigungucode?: string;
  tel?: string;
  title?: string;
  zipcode?: string;
  booktour?: string;
}